coco_segmentation/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
albumentations>=1.3.0
pycocotools>=2.0.6
pytorch-lightning>=2.0.0
tensorboard>=2.14.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0
